requirements.txt
astroid==1.4.8
beautifulsoup4==4.5.1
gitdb2==2.0.0
GitPython==2.1.0
isort==4.2.5
lazy-object-proxy==1.2.2
mccabe==0.5.2
peewee==2.8.5
perceval==0.4.0
pkg-resources==0.0.0
pylint==1.6.4
python-dateutil==2.6.0
requests==2.11.1
six==1.10.0
smmap2==2.0.1
wrapt==1.10.8
